RF2to3: working through futurize stage 2
This still includes a lot of old_div() calls that I'd like to get rid of

Also failing on some rendering tests

+++ psychopy/demos/coder/iohub/iosync/digitalInTest.py
@@ -42,7 +42,7 @@ getTime=core.getTime
 try:
     psychopy_mon_name='testMonitor'
     exp_code='events'
-    sess_code='S_{0}'.format(long(time.mktime(time.localtime())))
+    sess_code='S_{0}'.format(int(time.mktime(time.localtime())))
     iohub_config={
     "psychopy_monitor_name":psychopy_mon_name,
     "mcu.iosync.MCU":dict(serial_port='auto',monitor_event_types=['DigitalInputEvent',]),
